rename states -> state
