+++ lib/screens/main_screen/main_screen.dart
@@ -23,7 +23,7 @@ import 'package:picos/repository/medications_repository.dart';
 import 'package:picos/screens/add_medication_screen/add_medication_screen.dart';
 import 'package:picos/screens/login_screen.dart';
 import 'package:picos/screens/my_medications_screen/my_medications_screen.dart';
-import 'package:picos/states/medications_list_bloc.dart';
+import 'package:picos/state/medications_list_bloc.dart';
 import 'package:picos/themes/global_theme.dart';
 
 import '../../api/local_storage_medications_api.dart';
